Make tomviz a namespace package
The tomviz package is shared between the acquisition server and
the pipeline modules, which can be install independently.
