+++ tomviz/python/tomviz/__init__.py
@@ -1,5 +1,1 @@
-import os
-
-
-def in_application():
-    return os.environ.get('TOMVIZ_APPLICATION', False)
+__path__ = __import__('pkgutil').extend_path(__path__, __name__)
